Tidy CAR generated header files (#180)
Removed "_Module.h" and "Module.Export.h", only remain the "Module.h" file.
Use "static const UUID", removed "UUIDs.lib".
Defined macro _INSIDE_MODULE_ in some backgroup lube template.
Not using the "header.lub" template to generate header files.

Solved all the warnings in the devtools building environment.
Treat all the warnings as an error in the devtools building environment.
Default set to release version when entering the devtools building environment.

Moved syscar/eco/* to syscar/*
Added "MarshalRPCTest" and "HelloCar" executable to build.

Removed "xdk_gen_headers_from_cls" and added "xdk_export_car_headers" in
common.cmake.

+++ Sources/ToolChains/Carl/ElCar/Carc/clsbase/flat.cpp
@@ -909,7 +909,7 @@ int CFlatBuffer::CalcBufferSize(const CLSModule *pModule)
 
 size_t FlatCLS(const CLSModule *pSrc, void **ppDest)
 {
-    size_t nSize;
+    long nSize;
     CLSModule *pDest;
     CFlatBuffer buffer;
